% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignColumnNameVars.R
\name{assignColumnNameVars}
\alias{assignColumnNameVars}
\title{Assign Column Name Variables}
\usage{
assignColumnNameVars(prefixTuples, formulaTuples)
}
\arguments{
\item{prefixTuples}{See the return object from getProductionData.}

\item{formulaTuples}{See the return object from getProductionData.}
}
\value{
No value is returned, but many global variables are assigned.
}
\description{
This function acts as a helper function by assigning the column names to the
global environment.  This makes code much easier to read as it avoids the
use of continual get() statements within data.table.
}

