% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceAreaHarvested.R
\name{balanceAreaHarvested}
\alias{balanceAreaHarvested}
\title{Function to compute area harvested when new production and yield are given.}
\usage{
balanceAreaHarvested(data, processingParameters, formulaParameters)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{processingParameters}{A list of the parameters for the production
processing algorithms. See \code{productionProcessingParameters} for a
starting point.}

\item{formulaParameters}{A list holding the names and parmater of formulas.
See \code{productionFormulaParameters}.}
}
\description{
Function to compute area harvested when new production and yield are given.
}

