% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceProductionTriplet.R
\name{balanceProductionTriplet}
\alias{balanceProductionTriplet}
\title{This function balances the production triplet (input, productivity, output)
values for production where values are missing.}
\usage{
balanceProductionTriplet(data, processingParameters, formulaParameters)
}
\arguments{
\item{data}{The data containing the production triplet}

\item{processingParameters}{A list of the parameters for the production
processing algorithms.  See defaultProductionParameters() for a starting
point.}

\item{formulaParameters}{A list holding the names and parmater of formulas.
See \code{productionFormulaParameters}.}
}
\value{
Data where production triplet is calculated where available.
}
\description{
To perform this function, 0M values and previous values calculated which has
method flag 'i' should be removed as the function only fill in values which
has the value NA.
}

