% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateProcessingShare.R
\name{calculateProcessingShare}
\alias{calculateProcessingShare}
\title{Function to compute Processing sharing: share that identify the quantity of primary availability 
that it is allocated in different productive processes.}
\usage{
calculateProcessingShare(data, printSharesGraterThan1 = FALSE, param)
}
\arguments{
\item{data}{data table containing all the columns to compute processingSharing}

\item{params}{defaultProcessedItamParams parameters, object which contains the parameters}
}
\description{
Function to compute Processing sharing: share that identify the quantity of primary availability 
that it is allocated in different productive processes.
}

