% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAllProductionBalanced.R
\name{checkAllProductionBalanced}
\alias{checkAllProductionBalanced}
\title{Function to check whether the production itdentity is satisfied.}
\usage{
checkAllProductionBalanced(dataToBeSaved, formulaTable)
}
\arguments{
\item{dataToBeSaved}{The data.table to be saved back to the SWS.}

\item{formulaTable}{The table containing the formula constructed
by \code{constructFormulaTable}}
}
\value{
If the data satisfy the production identity, then the
    original data to be tested is returned. Otherwise, an error is
    raised.
}
\description{
This is a wrapper function for the function
\code{checkProductionBalanced} for multiple formula. Formula for
crops and meat corresponds to different elements, and this wrapper
enables one to check all the formulas specified in the formula
table.
}

