% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIdentityCalculated.R
\name{checkIdentityCalculated}
\alias{checkIdentityCalculated}
\title{Function to check whether the triplet (area harvested/production/yield) are
calculated whereever possible.}
\usage{
checkIdentityCalculated(dataToBeSaved, areaVar, yieldVar, prodVar,
  returnData = TRUE, normalised = TRUE)
}
\arguments{
\item{dataToBeSaved}{The data.table to be saved back to the SWS.}

\item{areaVar}{The column name corresponding to the area harvested.}

\item{yieldVar}{The column name corresponding to the yield.}

\item{prodVar}{The column name corresponding to produciton.}

\item{returnData}{logical, whether the data should be returned.}

\item{normalised}{logical, whether the data is normalised}
}
\value{
If the data satisfy the production identity, then the
    original data to be tested is returned. Otherwise, an error is
    raised.
}
\description{
Function to check whether the triplet (area harvested/production/yield) are
calculated whereever possible.
}

