% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProductionBalanced.R
\name{checkProductionBalanced}
\alias{checkProductionBalanced}
\title{Function to check whether the production itdentity is satisfied.}
\usage{
checkProductionBalanced(dataToBeSaved, areaVar, yieldVar, prodVar, conversion)
}
\arguments{
\item{dataToBeSaved}{The data.table to be saved back to the SWS.}

\item{formulaTable}{The table containing the formula constructed
by \code{constructFormulaTable}}
}
\value{
If the data satisfy the production identity, then the
    original data to be tested is returned. Otherwise, an error is
    raised.
}
\description{
Function to check whether the production itdentity is satisfied.
}

