% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Clean Production Data}
\usage{
cleanData(datasets, i, maxYear = NULL)
}
\arguments{
\item{datasets}{An object as produced by getProductionData.}

\item{i}{The row number of formulaTuples which we are currently working 
with.  At the time of writing this function, some commodities could have 
multiple different imputations (biological and indigineous meat is stored 
under the code for meat).  This is likely to be changed eventually, and 
this argument would then no longer be needed.}

\item{maxYear}{The final values in a time series can be blank cells, and this 
algorithm should provide estimates for such missing values.  As such, this
function creates 0Mu values up to the maximum year required.  This maximum
year is determined by either the maximum of the year variable in datasets 
(if this parameter is NULL) or by this parameter.  Note that setting a
maximum year that is smaller than the maximum year in the dataset will do
nothing.}
}
\value{
No object is returned, but the passed dataset is modified in place 
  via data.table.
}
\description{
This function does some basic cleaning of the dataset, such as removing 
unneeded columns and checking if there is any missing data to impute.
}

