% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineImputation.R
\name{combineImputation}
\alias{combineImputation}
\title{This function combines imputation from both model in which one can contain
estimated figures while the other does not.}
\usage{
combineImputation(imputationWithEstimates, imputationWithoutEstimates)
}
\arguments{
\item{imputationWithEstimates}{The imputed dataset where estimates are
included}

\item{imputationWithoutEstimates}{The imputed dataset where estimates were
removed.}
}
\value{
A combined dataset where rbind of both dataset
}
\description{
This function combines imputation from both model in which one can contain
estimated figures while the other does not.
}

