% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTotSlaughtered.R
\name{computeTotSlaughtered}
\alias{computeTotSlaughtered}
\title{computeTotStock}
\usage{
computeTotSlaughtered(data, tradeElements,
  FormulaParameters = animalFormulaParameters)
}
\arguments{
\item{data}{datataset containing the the numeber of animal stocks (trade not included)}
}
\value{
Returns a dataset with the the tot stock
}
\description{
This function has been created to compute the totStock that has to be used as
reference to compue the take off rates and consequently the number of slaughtered  animals.
In order to compute the total animal stocks, trade has to be taken into account
}

