% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFormulaTable.R
\name{constructFormulaTable}
\alias{constructFormulaTable}
\title{This function construct a formula table}
\usage{
constructFormulaTable(formulaTuples, formulaPrefix,
  whichPrefix = c("valuePrefix", "flagObsPrefix", "flagMethodPrefix"))
}
\arguments{
\item{formulaTuples}{The containing the formula relationship, this
can be obtained from the \code{getYieldFormula} function.}

\item{formulaPrefix}{This table contains the prefix which is used
to contruct the colnames. Can be obtained from the
\code{getFormulaPrefix} function.}

\item{whichPrefix}{This parameter determines which of the three
prefix in the formula prefix table should be concatenate in
the name.}
}
\value{
A formula table
}
\description{
The formula table is used by the \code{checkAllProductionBalanced}
function.
}

