% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createImputationObjectName.R
\name{createImputationObjectName}
\alias{createImputationObjectName}
\title{This function creates the name for the imputation object to be created and
loaded. This is to ensure the names are consistent between the two moduels.}
\usage{
createImputationObjectName(item)
}
\arguments{
\item{item}{The item code for the imputation (e.g. 0111 for wheat)}
}
\value{
A character name
}
\description{
This function creates the name for the imputation object to be created and
loaded. This is to ensure the names are consistent between the two moduels.
}

