% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultLogistic_DELETEME.R
\name{defaultLogistic}
\alias{defaultLogistic}
\title{Default Logistic Model for the Ensemble.}
\usage{
defaultLogistic(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\value{
A numeric vector with the estimated logistic regression model.
}
\description{
The model fit is \eqn{ Production = A/(1 + exp(-B(time-C))) }, and
fitting is done via non-linear least squares (see ?nls).  If this fit fails
to converge, than A is fixed to the maximum Production value and the model
is fit via glm().
}
\note{
If the midpoint of the logistic regression model is outside the range
of the data, then a vector of NA's is returned (to prevent poor
extrapolation).
}

