% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultProcessedItemParams.R
\name{defaultProcessedItemParams}
\alias{defaultProcessedItemParams}
\title{Default Processed Item Parameters}
\usage{
defaultProcessedItemParams()
}
\value{
A list with the Processed Item  parameters.
}
\description{
Provides an object which contains the parameters.  This 
allows for easy passing into functions.  The meaning of most variables is 
fairly clear, but a few obscure ones are described below:
}
\details{
\itemize{

\item processingShare share that identify the quantity of primary availability that it is allocated in 
different productive processes
\item shareDownUp Share to transform derived commodities into primary equivalent
\item availability The availability can be interpreted as the amount of one good that is no
 consumed as it is. This quantity enters, as input, in a productive process in order to produce
 derived commodities.

}
}

