% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureImputationInputs_DELETEME.R
\name{ensureImputationInputs}
\alias{ensureImputationInputs}
\title{Ensure Imputation Inputs}
\usage{
ensureImputationInputs(data, imputationParameters)
}
\arguments{
\item{data}{The data.frame containing the data to be imputed.}

\item{imputationParameters}{A list of the imputation parameters, such as
one generated by defaultImputationParameters().}
}
\value{
No value is returned, but errors are thrown if the input parameters
are not as expected.
}
\description{
Ensure Imputation Inputs
}

