% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureProductionInputs.R
\name{ensureProductionInputs}
\alias{ensureProductionInputs}
\title{Check Production Inputs}
\usage{
ensureProductionInputs(data, processingParameters, formulaParameters,
  returnData = TRUE, normalised = TRUE)
}
\arguments{
\item{data}{A data.table containing the data.}

\item{processingParameters}{A list containing the parameters to be used in
the processing algorithms.  See ?defaultProcessingParameters for a starting
point.}

\item{returnData}{logical, whether the data should be returned}

\item{normalised}{logical, whether the data is normalised}
}
\value{
The original data if all tests are passed
}
\description{
This function is designed to ensure that the provided dataset is valid.
}

