% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureProductionOutputs.R
\name{ensureProductionOutputs}
\alias{ensureProductionOutputs}
\title{Check Production Inputs}
\usage{
ensureProductionOutputs(data, processingParameters, formulaParameters,
  testImputed = TRUE, testCalculated = TRUE, domain = "agriculture",
  dataset = "aproduction", returnData = TRUE, normalised = TRUE)
}
\arguments{
\item{data}{A data.table containing the data.}

\item{processingParameters}{A list containing the parameters to be used in
the processing algorithms. See ?defaultProcessingParameters for a
starting point.}

\item{formulaParameters}{A list holding the names and parmater of formulas.
See \code{productionFormulaParameters}.}

\item{testImputed}{logical, whether to test imputation result}

\item{testCalculated}{logical, whether to test calculated result.}

\item{domain}{The domain to be saved back}

\item{dataset}{The dataset to be saved back}

\item{returnData}{logical, whether the data should be returned}

\item{normalised}{logical, whether the data is normalised}
}
\value{
The original data if all tests are passed
}
\description{
This function is designed to ensure that the dataset to be saved is valid.
}

