% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandDatasetYears.R
\name{expandDatasetYears}
\alias{expandDatasetYears}
\title{This function guarantees that for the years for which imputations are requested
there are empty cells flaged as (M, u) to be overwritten with the imputed fugures.}
\usage{
expandDatasetYears(data, params, startYear, endYear)
}
\arguments{
\item{data}{The set of data for which we want to expand the year}

\item{params}{List}

\item{startYear}{ffkf}

\item{endYear}{rrrrrrr}
}
\value{
test duddudu
}
\description{
This function guarantees that for the years for which imputations are requested
there are empty cells flaged as (M, u) to be overwritten with the imputed fugures.
}

