% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandMeatSessionSelection.R
\name{expandMeatSessionSelection}
\alias{expandMeatSessionSelection}
\title{This function expands the current selection in the session to
include all necessary meat.}
\usage{
expandMeatSessionSelection(oldKey, selectedMeatTable,
  areaVar = "geographicAreaM49", itemVar = "measuredItemCPC",
  elementVar = "measuredElement", yearVar = "timePointYears")
}
\arguments{
\item{oldKey}{The current key used in the session}

\item{selectedMeatTable}{The data table of the same format returned by
the function \code{getAnimalMeatMapping}, but containing the
meat selection required.}
}
\value{
A new DatasetKey object expanded to include the meat
    selection specified.
}
\description{
This function expands the current selection in the session to
include all necessary meat.
}

