% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillRecord.R
\name{fillRecord}
\alias{fillRecord}
\title{This function replaces non-existing records with missing records.}
\usage{
fillRecord(data, valuePattern = "Value",
  flagObsPattern = "flagObservationStatus",
  flagMethodPattern = "flagMethod")
}
\arguments{
\item{data}{The dataset}

\item{valuePattern}{The regular expression to capture the value column.}

\item{flagObsPattern}{The regular expression to capture the observation
status flag column.}

\item{flagMethodPattern}{The regular expression to capture the method flag
column.}
}
\value{
dataset with identical dimension but creates missing record in place
    of non-existing record.
}
\description{
Occastionally the database may return records where all three columns
(value, observation flag, method flag) are all NA and is a non-existing
record in the database; this can also be created when denormalising data. We
will replace is with a missing record representation where (Value = NA,
flagObservationStatus = "M" and flagMethod = "u").
}

