% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTimeRange.R
\name{filterTimeRange}
\alias{filterTimeRange}
\title{This functino filters the time dimension of a data}
\usage{
filterTimeRange(data, firstYear, lastYear)
}
\arguments{
\item{data}{The data table object}

\item{firstYear}{The starting year of the filtering}

\item{lastYear}{The last year of the filtering}
}
\value{
A filtered dataset
}
\description{
This functino filters the time dimension of a data
}

