% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllAnimalNumber.R
\name{getAllAnimalNumber}
\alias{getAllAnimalNumber}
\title{This function retrieves all the animal number of the animal
commodity in the data base.}
\usage{
getAllAnimalNumber(animalMeatMapping = getAnimalMeatMapping())
}
\arguments{
\item{animalMeatMapping}{The mapping table which contains the
relationship between the animal commodity and the meat
commodity. Can be retrieved from \code{getAnimalMeatMapping}
function.}
}
\value{
A data.table containing animal numbers
}
\description{
This function retrieves all the animal number of the animal
commodity in the data base.
}

