% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAllHistory.R
\name{getAllHistory}
\alias{getAllHistory}
\title{Get All Production History}
\usage{
getAllHistory()
}
\value{
A data.table object with the following columns:
\itemize{
    \item geographicAreaM49: The code for the country of this observation.
    \item measuredElement: The code for the element of this observation
    (these match with the "Element" dimension in the agriculture domain,
    for example 5312 is area harvested, 5510 is production, etc.).
    \item measuredItemCPC: The CPC code for the commodity.
    \item timePointYears: The year for this observation.
    \item Version: ?
    \item StartDate: The initial date this value was entered into the
    system.
    \item EndDate: The date that this value was overwritten.  If NA, this
    value is the current best estimate and is considered valid.  If not NA,
    this value is considered invalid.
    \item Metadata: ?
    \item Metadata_Language: An abbreviation for the language of the
    metadata comments.
    \item Metadata_Group: ?
    \item Metadata_Element: ?
    \item Metadata_Value: ?
    \item Value: The value stored in the database for this observation.
    \item flagObservationStatus: The flag corresponding to the observation
    status of the current observation.  See the faoswsFlag package for more
    details.
    \item flagMethod: The method flag for the current observation.  Again,
    see the faoswsFlag package.
    \item valid: A Y/N value indicating whether or not the current value is
    valid.  This is determined solely on the basis of whether or not this
    value was overwritten.
    \item productionValue: The value corresponding to production (element
    5510) for this particular observation.  This may be useful in training
    the classification ensemble.
}
}
\description{
This function is used to pull the historical production values from the
database.  One main purpose of this data is to train machine learning
algorithms on invalid observations, and use those models to attempt to
detect invalid observations in the current data.
}

