% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllSlaughteredNumber.R
\name{getAllSlaughteredNumber}
\alias{getAllSlaughteredNumber}
\title{This function retrieves all the animal slaughtered number of the
meat commodity in the data base.}
\usage{
getAllSlaughteredNumber(animalMeatMapping = getAnimalMeatMapping())
}
\arguments{
\item{animalMeatMapping}{The mapping table which contains the
relationship between the animal commodity and the meat
commodity. Can be retrieved from \code{getAnimalMeatMapping}
function.}
}
\value{
A data.table containing animal slaughtered numbers
}
\description{
This function retrieves all the animal slaughtered number of the
meat commodity in the data base.
}

