% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnimalMeatMapping.R
\name{getAnimalMeatMapping}
\alias{getAnimalMeatMapping}
\title{This function retrieves the mapping between the parent (animal
such as cattle) and child (meat such as meat of cattle).}
\usage{
getAnimalMeatMapping(R_SWS_SHARE_PATH, onlyMeatChildren = FALSE,
  meatPattern = "^211(1|2|7).*")
}
\arguments{
\item{R_SWS_SHARE_PATH}{The path to the shared drive on the SWS.}

\item{onlyMeatChildren}{Whether only the meat children should be returned.}

\item{meatPattern}{The regular expression corresponding to meat}
}
\value{
The mapping table
}
\description{
This function retrieves the mapping between the parent (animal
such as cattle) and child (meat such as meat of cattle).
}

