% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompleteImputationKey.R
\name{getCompleteImputationKey}
\alias{getCompleteImputationKey}
\title{This function reads the complete imputation key from the Datatable
'fbs_production_comm_codes'}
\usage{
getCompleteImputationKey()
}
\value{
A DatasetKey object containing all the keys for production
    imputation
}
\description{
This function reads the complete imputation key from the Datatable
'fbs_production_comm_codes'
}

