% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getForcedZeroKey.R
\name{getForcedZeroKey}
\alias{getForcedZeroKey}
\title{This function determine which observations shouldn't be imputed:
Production should not be imputed on 0Mn observations.  These are
"missing but assumed negligble."  Additionally, we have not been
able to identify 0M from the old system as 0Mu or 0Mn and have
thus assigned them the flags 0M-.  These should be treated as 0Mn
in this case.}
\usage{
getForcedZeroKey(data, processingParams = defaultProcessingParameters(),
  productionParams = getImputationParameters()$productionParams)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{productionParams}{The production imputation parameter
returned by the function \code{getImputationParameters}.}

\item{processingParameters}{A list of the parameters for the
production processing algorithms. See
\code{defaultProcessingParameters} for a starting point.}
}
\value{
A data.table containing the keys where the value should be
    forced to be zero.
}
\description{
This function determine which observations shouldn't be imputed:
Production should not be imputed on 0Mn observations.  These are
"missing but assumed negligble."  Additionally, we have not been
able to identify 0M from the old system as 0Mu or 0Mn and have
thus assigned them the flags 0M-.  These should be treated as 0Mn
in this case.
}

