% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFormulaPrefix.R
\name{getFormulaPrefix}
\alias{getFormulaPrefix}
\title{This function returns the prefix name used for the denormalised
dataset.}
\usage{
getFormulaPrefix()
}
\value{
A data.table containing the prefix
}
\description{
TODO (Michael): This function should probably be merged with
                processing parameters.
}

