% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImputedValues.R
\name{getImputedValues}
\alias{getImputedValues}
\title{This function extracts the imputation depending on whether estimates are
used}
\usage{
getImputedValues(data, useEstimatesTable, useEstimates)
}
\arguments{
\item{data}{The imputed dataset}

\item{useEstimatesTable}{The table created using the function
\code{useEstimateForTimeSeriesImputation}.}

\item{useEstimates}{logical, whether estimates should be used.}
}
\description{
This function extracts the imputation depending on whether estimates are
used
}

