% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastValue.R
\name{getLastValue}
\alias{getLastValue}
\title{Get Last Value}
\usage{
getLastValue(d, missingObsFlag = "M", missingMetFlag = "u")
}
\arguments{
\item{missingObsFlag}{The observation flag for missing observations.}

\item{missingMetFlag}{The method flag for missing observations.}

\item{A}{data.table with columns as described above.}
}
\value{
Nothing is returned, but the passed dataset is filtered to the
  correct rows.
}
\description{
Get Last Value
}

