% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMainKey.R
\name{getMainKey}
\alias{getMainKey}
\title{Get Main Key}
\usage{
getMainKey(years)
}
\arguments{
\item{years}{A numeric vector of the years which should be extracted and
used in this analysis.}
}
\value{
A DatasetKey object.
}
\description{
This function generates a DatasetKey object (see faosws definition) which is
used to query the production dataset.
}

