% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrimary.R
\name{getPrimary}
\alias{getPrimary}
\title{Function trace back to the primary item starting from an edge-item.
It uses the commodity tree to go back level by level untill the first one.}
\usage{
getPrimary(edge, tree, standParams)
}
\arguments{
\item{edge}{a CPC (or a vector of CPC codes)}

\item{tree}{Commodity tree, data.table}

\item{standParams}{computation parameters}
}
\description{
Function trace back to the primary item starting from an edge-item.
It uses the commodity tree to go back level by level untill the first one.
}

