% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getProductionData.R
\name{getProductionData}
\alias{getProductionData}
\title{Get Production Data}
\usage{
getProductionData(dataContext)
}
\arguments{
\item{dataContext}{An object of type DatasetKey (defined in faosws).  This
parameter provides the configuration of the dataset (i.e. which countries,
commodities, etc.).}
}
\value{
A list of length 3:
\itemize{
\item query: A data.table containing the production data from the database.
\item formulaTuples: The element codes for the production, output, and yield
elements.  There may be multiple records in this table if multiple elements
are relevant.
\item prefixTuples: This element contains the prefixes for the flags and
values.
}
}
\description{
This function returns the production data currently in the database for a
particular DatasetKey.  It also performs a few clean-up operations on the
data before directly returning.
}

