% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProductionFormula.R
\name{getProductionFormula}
\alias{getProductionFormula}
\title{Get Yield Formula}
\usage{
getProductionFormula(itemCode, itemVar = "measuredItemCPC", warn = FALSE)
}
\arguments{
\item{itemCode}{The cpc code providing the commodity of interest.  This
should be a character string.}

\item{itemVar}{The name of the item variable column.}

\item{warn}{Logical.  If FALSE, an error is thrown for missing element
types.  If TRUE, it is simply a warning.}
}
\value{
A data.table object providing the relevant element codes.  In some
  cases, there may be multiple rows, as it is possible to have multiple
  elements which hold production, yield, and output data.
}
\description{
This function provides the element codes (for production, yield, and output)
and conversion factor which correspond to the particular item code of
interest.
}

