% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSessionMeatSelection.R
\name{getSessionMeatSelection}
\alias{getSessionMeatSelection}
\title{This function filters the selected keys to the selected meats}
\usage{
getSessionMeatSelection(key, selectedMeatTable)
}
\arguments{
\item{key}{The DatasetKey object of the current session}

\item{selectedMeatTable}{The table returned by the function
\code{getAnimalMeatMapping}.}
}
\value{
A character list containing the required meat commodity code.
}
\description{
This function filters the selected keys to the selected meats
}

