% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSlaughteredAnimal.R
\name{getSlaughteredAnimal}
\alias{getSlaughteredAnimal}
\title{Function to delete all other elements but only retain elements corresponding
to slaughtered animal}
\usage{
getSlaughteredAnimal(data, formulaTuples)
}
\arguments{
\item{data}{The dataset}

\item{formulaTuples}{The output of \code{getYieldFormula}}
}
\value{
A data table where non slaughtered animal elements are removed
}
\description{
Function to delete all other elements but only retain elements corresponding
to slaughtered animal
}

