% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSUADataRestricted.R
\name{getSUADataRestricted}
\alias{getSUADataRestricted}
\title{This function pull the SUA table necessary to compute the availabilities
for computing the processed item productions.}
\usage{
getSUADataRestricted()
}
\description{
This function pull the SUA table necessary to compute the availabilities
for computing the processed item productions.
}

