% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTree.R
\name{getTree}
\alias{getTree}
\title{GetTree}
\usage{
getTree()
}
\arguments{
\item{data}{input data.table: SUA table}

\item{tree}{Commodity tree, data.table}

\item{printNegativeAvailability}{It is TRUE if you want to produce csv file containing nagative availabilities}

\item{params}{computation parameters}
}
\description{
This function has been created in order to pull the commodity tree from the SWS.
Despite such kind of function already exists (see getcommodityTree), we created this 
additional function in order to modify/simplufy the commodity for
the processed items production imputation process. In particular we need to remove some   
connections, add some extraction rates and add the columns containing the processing level.
}

