% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYieldData.R
\name{getYieldData}
\alias{getYieldData}
\title{This function extracts the yield data}
\usage{
getYieldData(dataContext)
}
\arguments{
\item{dataContect}{The data contect returned by \code{GetTestEnvironment}}
}
\value{
A list containing the yield data, the yield formula and also the
    prefix of the column names.
}
\description{
This function extracts the yield data
}

