% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getYieldFormula.R
\name{getYieldFormula}
\alias{getYieldFormula}
\title{Get Yield Formula}
\usage{
getYieldFormula(itemCode)
}
\arguments{
\item{itemCode}{The cpc code providing the commodity of interest.  This
should be a character string.}
}
\value{
A data.table object providing the relevant element codes.  In some
cases, there may be multiple rows, as it is possible to have multiple
elements which hold production, yield, and output data.
}
\description{
This function provides the element codes (for production, yield, and output)
and conversion factor which correspond to the particular item code of
interest.
}

