% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationExist.R
\name{imputationExist}
\alias{imputationExist}
\title{Check whether the imputation model exist}
\usage{
imputationExist(modelLoadingPath, item)
}
\arguments{
\item{modelLoadingPath}{The path to where imputed datasets are placed.}

\item{item}{The commodity item code}
}
\value{
logical, whether the specific model exist
}
\description{
Check whether the imputation model exist
}

