% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMeatTriplet.R
\name{imputeMeatTriplet}
\alias{imputeMeatTriplet}
\title{This is function performs imputation on the triplet element. However, this
is an implementation from Josh.}
\usage{
imputeMeatTriplet(meatKey, minObsForEst = 5)
}
\arguments{
\item{meatKey}{The Datakey object}

\item{minObsForEst}{The minimum number of observation required to keep
'E'stimated values.}
}
\value{
Imputed time series
}
\description{
NOTE (Michael); This function should ultimately be removed, it contains a
lot og hacks is unreasonable implementations.
}

