% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeProductionDomain.R
\name{imputeProductionDomain}
\alias{imputeProductionDomain}
\title{This function imputes the whole production domain.}
\usage{
imputeProductionDomain(data, processingParameters,
  areaHarvestedImputationParameters, yieldImputationParameters,
  productionImputationParameters, unitConversion)
}
\arguments{
\item{data}{The data}

\item{processingParameters}{A list of the parameters for the production
processing algorithms. See defaultProductionParameters() for a starting
point.}

\item{areaHarvestedImputationParameters}{A list of the parameters for the
area harvested imputation. See defaultImputationParameters() for a
starting point.}

\item{yieldImputationParameters}{A list of the parameters for the yield
imputation. See defaultImputationParameters() for a starting point.}

\item{productionImputationParameters}{A list of the parameters for the
production imputation. See defaultImputationParameters() for a starting
point.}

\item{unitConversion}{The multiplicative factor used for the particular
commodity in calculating the yield.}
}
\description{
The function will impute production, area harvested and yield at the same
time.
}
\details{
Transformation in the yield formula is not allowed and will not be taken
into account.
}

