% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeProductionTriplet.R
\name{imputeProductionTriplet}
\alias{imputeProductionTriplet}
\title{This function imputes the whole production domain.}
\usage{
imputeProductionTriplet(data, processingParameters, formulaParameters,
  imputationParameters)
}
\arguments{
\item{data}{The data}

\item{processingParameters}{A list of the parameters for the production
processing algorithms. See defaultProductionParameters() for a starting
point.}

\item{formulaParameters}{A list holding the names and parmater of formulas.
See \code{productionFormulaParameters}.}

\item{imputationParameters}{A list holding the imputation parameters, see
\code{getImputationParameters}.}
}
\description{
The function will impute production, area harvested and yield at the same
time.
}
\details{
Transformation in the yield formula is not allowed and will not be taken
into account.
}

