% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPrimary.R
\name{isPrimary}
\alias{isPrimary}
\title{This function tests whether a commodity is a primary item.}
\usage{
isPrimary(itemCode, primaryPattern = "^0")
}
\arguments{
\item{itemCode}{The item code to be tested}

\item{primaryPattern}{The regular expression pattern corresponds to primary
commodities.}
}
\value{
logical, whether the item is a primary commodity
}
\description{
HACK (Michael): This is a hack from Josh, a list should be provided.
}

