% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonImputationItems.R
\name{nonImputationItems}
\alias{nonImputationItems}
\title{This function returns the commodity codes which are selected in the session
but not in the imputation list}
\usage{
nonImputationItems(selectedKey, imputationKey)
}
\arguments{
\item{selectedKey}{Datakey of the current session.}

\item{imputationKey}{The object returned by the function
\code{getImputationKey}}
}
\value{
A vector containing the commodity items that are selected but not
    included in the imputation list.
}
\description{
The purpose of the function to capture and give warning when items are being
selected in the session, but not included in the imputation list.
}

