% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResult.R
\name{plotResult}
\alias{plotResult}
\title{This function has been created to plot PROCESSED data results
It is very Module-specific}
\usage{
plotResult(data, toPubblish, pathToSave)
}
\arguments{
\item{toPubblish}{}

\item{pathToSave}{}
}
\description{
This function has been created to plot PROCESSED data results
It is very Module-specific
}

