% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productionFormulaParameters.R
\name{productionFormulaParameters}
\alias{productionFormulaParameters}
\title{Parameter for formulas}
\usage{
productionFormulaParameters(datasetConfig, productionCode = "5510",
  areaHarvestedCode = "5312", yieldCode = "5416", unitConversion = 1)
}
\arguments{
\item{datasetConfig}{The dataset configuration returned by the function
\code{GetDatasetConfig}}

\item{productionCode}{The element code corresponding to the production
variable.}

\item{areaHarvestedCode}{The element code corresponding to the area
harvested variable.}

\item{yieldCode}{The element code corresponding to the yield variable.}

\item{unitConversion}{The unit conversion between the triplet code provided.}
}
\value{
Returns a list of the formula parameters used in the data
  pre-processing algorithm.
}
\description{
Parameter for formulas
}
\details{
Below is a description of the parameters: \itemize{
\item productionValue: The column name of the production variable.
\item productionObservationFlag: The column name of the observation flag
  corresponding to the production variable.
\item productionMethodFlag: The column name of the method flag corresponding
  to the production variable.
\item yieldValue: The column name of the yield variable.
\item unitConversion: The unit conversion between the triplet code provided.
\item yieldObservationFlag: The column name of the observation flag
  corresponding to the yield variable.
\item yieldMethodFlag: The column
  name of the method flag corresponding to the yield variable.
\item areaHarvestedValue: The column name of the area harvested variable.
\item areaHarvestedObservationFlag: The column name of the observation flag
  corresponding to the area harvested variable.
\item areaHarvestedMethodFlag: The column name of the method flag
  corresponding to the area harvested variable.
\item unitConversion The unit conversion applied in the formula
}
}

