% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeForcedZero.R
\name{removeForcedZero}
\alias{removeForcedZero}
\title{This function removes removes values that should be forced zero.}
\usage{
removeForcedZero(data, forcedZeroKey)
}
\arguments{
\item{data}{The dataset to be filtered}

\item{forcedZeroKey}{The index returned by the function
\code{getForcedZeroKey}}
}
\value{
The dataset without index where data are forced to be zero
}
\description{
This function removes removes values that should be forced zero.
}

