% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeIndigenousBiologicalMeat.R
\name{removeIndigenousBiologicalMeat}
\alias{removeIndigenousBiologicalMeat}
\title{This function removes the entry which corresponds to indigenous and
biological meats.}
\usage{
removeIndigenousBiologicalMeat(formula)
}
\arguments{
\item{formula}{The object returned by the \code{getYieldFormula}}
}
\value{
The filtered formula without biological and indigenous treatment.
}
\description{
This function removes the entry which corresponds to indigenous and
biological meats.
}

