% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMissingEntry.R
\name{removeMissingEntry}
\alias{removeMissingEntry}
\title{This function filters data to remove missing entries.}
\usage{
removeMissingEntry(data)
}
\arguments{
\item{data}{The dataset}
}
\value{
A data table where flags with 'M' are removed.
}
\description{
This function filters data to remove missing entries.
}

