% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundResults.R
\name{roundResults}
\alias{roundResults}
\title{Round Results}
\usage{
roundResults(x)
}
\arguments{
\item{x}{A numeric value to be rounded.}
}
\value{
A numeric value after the rounding has occurred.
}
\description{
This function is used to round the production and area harvested values in
the production imputation and yield computation modules.
}

