% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveProductionData.R
\name{saveProductionData}
\alias{saveProductionData}
\title{Save Production Data}
\usage{
saveProductionData(data, areaHarvestedCode = "5312", yieldCode = "5421",
  productionCode = "5510", verbose = FALSE)
}
\arguments{
\item{data}{The data.table object containing the seed data to be written to
the database.}

\item{areaHarvestedCode}{Character string containing the element code
corresponding to the area harvested variable.}

\item{yieldCode}{Character string containing the element code
corresponding to the yield variable.}

\item{productionCode}{Character string containing the element code
corresponding to the production variable.}

\item{verbose}{Should output be printed about the progress of the save
data?  Defaults to FALSE.}
}
\value{
No R objects are returned, as this functions purpose is solely to
write to the database.
}
\description{
This function takes the a seed data dataset and saves it back to the
database.
}

