% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectImputationItem.R
\name{selectImputationItem}
\alias{selectImputationItem}
\title{This function returns the commodity codes which are selected in the session
and is in the imputation list}
\usage{
selectImputationItem(selectedKey, imputationKey)
}
\arguments{
\item{selectedKey}{Datakey of the current session.}

\item{imputationKey}{The object returned by the function
\code{getImputationKey}}
}
\value{
A vector containing the commodity items that are selected and is
    included in the imputation list.
}
\description{
This function returns the commodity codes which are selected in the session
and is in the imputation list
}

