% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectMeatCodes.R
\name{selectMeatCodes}
\alias{selectMeatCodes}
\title{This function selects the code which are corresponds to meat}
\usage{
selectMeatCodes(itemCodes, meatPattern = "^211(1|2|7).*")
}
\arguments{
\item{itemCodes}{Vector of item codes}

\item{meatPattern}{The regular expression which corresponds to meat items}
}
\value{
Item codes which corresponds to meat
}
\description{
This function selects the code which are corresponds to meat
}

