% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectParentData.R
\name{selectParentData}
\alias{selectParentData}
\title{This function subsets the complete data to include only commodities which
are considered a parent.}
\usage{
selectParentData(data, selectedMeatTable, itemVar = "measuredItemCPC",
  elementVar = "measuredElement")
}
\arguments{
\item{data}{The dataset}

\item{selectedMeatTable}{Table of the same format as the output of
\code{getAnimalMeatMapping}}

\item{itemVar}{The column name corresponding item commodity.}

\item{elementVar}{The column name corresponding to the measured element.}
}
\value{
A dataset where all commodities are parent.
}
\description{
In this case, the animal is the parent of meats.
}

