% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectUseEstimate.R
\name{selectUseEstimate}
\alias{selectUseEstimate}
\title{Subset imputed data to whether estimates should be used.}
\usage{
selectUseEstimate(data, useEstimatesTable, useEstimates)
}
\arguments{
\item{data}{The imputed dataset}

\item{useEstimatesTable}{The table returned by the function
\code{useEstimateForTimeSeriesImputation}}

\item{useEstimates}{Whether the estimates should be used.}
}
\value{
The subset of the imputed data
}
\description{
Subset imputed data to whether estimates should be used.
}

