% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetAnimalMeatMapping.R
\name{subsetAnimalMeatMapping}
\alias{subsetAnimalMeatMapping}
\title{This function subsets the animal meat mapping returned by the
\code{getAnimalMeatMapping} so that it contains item only in the
selected session.}
\usage{
subsetAnimalMeatMapping(animalMeatMapping, context)
}
\arguments{
\item{animalMeatMapping}{The mapping returned by the
\code{getAnimalMeatMapping}}

\item{context}{The swsContext.datasets returned by the
\code{GetTestEnvironment}.}
}
\value{
The same animal meat mapping table but with only items
    selected by the session.
}
\description{
This function subsets the animal meat mapping returned by the
\code{getAnimalMeatMapping} so that it contains item only in the
selected session.
}

