% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transferAnimalSlaughtered.R
\name{transferAnimalSlaughtered}
\alias{transferAnimalSlaughtered}
\title{Function to transfer animal slaughtered from animal to meat or back from
meat to animals.}
\usage{
transferAnimalSlaughtered(meatData, animalData, parentToChild = TRUE)
}
\arguments{
\item{meatData}{The animal slaughtered data from meat commodity.}

\item{animalData}{The animal slaughtered data from animal commodity.}

\item{parentToChild}{logical, if true, slaughtered animal are transferred
from animal commodity to meat, otherwise the otherway around.}
}
\value{
The transferred data
}
\description{
Function to transfer animal slaughtered from animal to meat or back from
meat to animals.
}

