% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transferParentToChild.R
\name{transferParentToChild}
\alias{transferParentToChild}
\title{The function transfer values from parent to child}
\usage{
transferParentToChild(parentData, commodityTree, selectedMeatTable)
}
\arguments{
\item{parentData}{The data containing the values of the parent}

\item{commodityTree}{The commodity tree returned by \code{getCommodityTree}.}

\item{selectedMeatTable}{Table of the same format as the output of
\code{getAnimalMeatMapping}.}
}
\value{
A dataset with the codes of children items while values updated from
    parents.
}
\description{
The function transfer values from parent to child
}

