% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateOriginalDataWithImputation.R
\name{updateOriginalDataWithImputation}
\alias{updateOriginalDataWithImputation}
\title{The function updates the unimputed data with imputed values}
\usage{
updateOriginalDataWithImputation(originalData, imputedData, yieldElementCode,
  prodElementCode)
}
\arguments{
\item{originalData}{The unimputed data}

\item{imputedData}{The imputed data}

\item{yieldElementCode}{The element code of corresponding to yield.}

\item{prodElementCode}{The element code of corresponding to production.}
}
\value{
An updated data
}
\description{
The function updates the unimputed data with imputed values
}

