% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceAreaHarvested.R
\name{balanceAreaHarvested}
\alias{balanceAreaHarvested}
\title{Function to compute area harvested when new production and yield are given.}
\usage{
balanceAreaHarvested(data, processingParameters, newMethodFlag = "i",
  unitConversion = 1)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{processingParameters}{A list of the parameters for the production
processing algorithms.  See defaultProductionParameters() for a starting
point.}

\item{newMethodFlag}{The flag which should be placed for computed
observations as the method flag.}

\item{unitConversion}{Yield is computed as (production) / (area) and
multiplied by unitConversion.  This parameter defaults to 1.}
}
\description{
Function to compute area harvested when new production and yield are given.
}

