% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/balanceProduction.R
\name{balanceProduction}
\alias{balanceProduction}
\title{Function to compute production when new area harvested and yield
are given.}
\usage{
balanceProduction(data, imputationParameters, processingParameters)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{imputationParameters}{A list of the parameters for the imputation
algorithms.  See defaultImputationParameters() for a starting point.}

\item{processingParameters}{A list of the parameters for the production
processing algorithms.  See defaultProductionParameters() for a starting
point.}
}
\description{
Function to compute production when new area harvested and yield
are given.
}

