% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProductionBalanced.R
\name{checkProductionBalanced}
\alias{checkProductionBalanced}
\title{Function to check whether the production itdentity is satisfied.}
\usage{
checkProductionBalanced(dataToBeSaved, areaVar, yieldVar, prodVar, conversion,
  returnData = TRUE, normalised = TRUE)
}
\arguments{
\item{dataToBeSaved}{The data.table to be saved back to the SWS.}

\item{areaVar}{The column name corresponding to the area harvested.}

\item{yieldVar}{The column name corresponding to the yield.}

\item{prodVar}{The column name corresponding to produciton.}

\item{conversion}{The conversion factor for calculating production.}

\item{returnData}{logical, whether the data should be returned.}

\item{normalised}{logical, whether the data is normalised}
}
\value{
If the data satisfy the production identity, then the
    original data to be tested is returned. Otherwise, an error is
    raised.
}
\description{
Function to check whether the production itdentity is satisfied.
}

