% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeYield.R
\name{computeYield}
\alias{computeYield}
\title{Function to compute and update yield}
\usage{
computeYield(data, processingParameters, normalized = FALSE,
  newMethodFlag = "i", flagTable = faoswsFlagTable, unitConversion = 1)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{processingParameters}{A list of the parameters for the production
processing algorithms.  See defaultProductionParameters() for a starting
point.}

\item{normalized}{Is the input data normalised.}

\item{newMethodFlag}{The flag to be used to update the yield method flag
when imputation occurs.}

\item{flagTable}{see data(faoswsFlagTable) in \pkg{faoswsFlag}}

\item{unitConversion}{Yield is computed as (production) / (area) and
multiplied by unitConversion.  This parameter defaults to 1.}
}
\value{
The updated data.table.  This is important in the case where the
  data is normalized, as the data.table must be cast and reshaped (and thus
  can't be modified by reference).
}
\description{
Function to compute and update yield
}

