% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeYield.R
\name{computeYield}
\alias{computeYield}
\title{Function to compute and update yield}
\usage{
computeYield(data, processingParameters, flagTable = faoswsFlagTable)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{processingParameters}{A list of the parameters for the production
processing algorithms.  See defaultProductionParameters() for a starting
point.}

\item{flagTable}{see data(faoswsFlagTable) in \pkg{faoswsFlag}}
}
\value{
The updated data.table.
}
\description{
Function to compute and update yield
}

