% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandMeatSessionSelection.R
\name{expandMeatSessionSelection}
\alias{expandMeatSessionSelection}
\title{This function expands the current selection in the session to
include all necessary meat.}
\usage{
expandMeatSessionSelection(oldKey, selectedMeatTable,
  areaVar = "geographicAreaM49", itemVar = "measuredItemCPC",
  elementVar = "measuredElement", yearVar = "timePointYears")
}
\arguments{
\item{oldKey}{The current key used in the session}

\item{selectedMeatTable}{animalMeatMapping The mapping returned by the
\code{getAnimalMeatMapping}}

\item{areaVar}{The column name corresponding to the geographic area.}

\item{itemVar}{The column name corresponding to the commodity item.}

\item{elementVar}{The column name corresponding to the measured element.}

\item{yearVar}{The column name corresponding to the time dimension.}
}
\value{
A new DatasetKey object expanded to include the meat
    selection specified.
}
\description{
This function expands the current selection in the session to
include all necessary meat.
}

