% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImputationParameters.R
\name{getImputationParameters}
\alias{getImputationParameters}
\title{Get Imputation Parameters}
\usage{
getImputationParameters(datasets)
}
\arguments{
\item{datasets}{The object being imputed on, typically created by
getProductionData.}
}
\value{
A list with the yield and production imputation parameters.
}
\description{
This function contains the logic that we'll use to define the imputation 
models and their parameters.
}

