% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImputationParameters.R
\name{getImputationParameters}
\alias{getImputationParameters}
\title{Get Imputation Parameters}
\usage{
getImputationParameters(datasets, i = 1, areaVar = "geographicAreaM49",
  itemVar = "measuredItemCPC", elementVar = "measuredElement",
  yearVar = "timePointYears")
}
\arguments{
\item{datasets}{The object being imputed on, typically created by
getProductionData.}

\item{i}{The row number of formulaTuples.  This allows R to produce the
correct formula tuple/yield codes, and is typically 1.}

\item{areaVar}{The column name corresponding to the geographic area.}

\item{itemVar}{The column name corresponding to the commodity item.}

\item{elementVar}{The column name corresponding to the measured element.}

\item{yearVar}{The column name corresponding to the time dimension}
}
\value{
A list with the yield and production imputation parameters.
}
\description{
This function contains the logic that we'll use to define the imputation
models and their parameters.
}

