% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imputeProductionDomain.R
\name{imputeProductionDomain}
\alias{imputeProductionDomain}
\title{This function imputes the whole production domain.}
\usage{
imputeProductionDomain(data, processingParameters, yieldImputationParameters,
  productionImputationParameters)
}
\arguments{
\item{data}{The data}

\item{processingParameters}{A list of the parameters for the production
processing algorithms.  See defaultProductionParameters() for a starting
point.}

\item{yieldImputationParameters}{A list of the parameters for the
yield imputation.  See defaultImputationParameters() for a starting point.}

\item{productionImputationParameters}{A list of the parameters for the
production imputation.  See defaultImputationParameters() for a starting
point.}
}
\description{
The function will impute production, area harvested and yield at
the same time.
}
\details{
Transformation in the yield formula is not allowed and will not be
taken into account.
}

