% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeWithAndWithoutEstimates.R
\name{imputeWithAndWithoutEstimates}
\alias{imputeWithAndWithoutEstimates}
\title{This is function performs imputation on the triplet element. This is an
implementation from Josh to account to exclude estimates when sufficient
data is available or use estimates when they are not.}
\usage{
imputeWithAndWithoutEstimates(meatKey, minObsForEst = 5)
}
\arguments{
\item{meatKey}{The Datakey object}

\item{minObsForEst}{The minimum number of observation required to keep
'E'stimated values.}
}
\value{
Imputed time series
}
\description{
NOTE (Michael): This function should ultimately be removed, it contains a
                lot of hacks and manual estimated values with method flag
                'f' should not be imputed.
}

