% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeWithAndWithoutEstimates.R
\name{imputeWithAndWithoutEstimates}
\alias{imputeWithAndWithoutEstimates}
\title{This is function performs imputation on the triplet element. This is an
implementation from Josh to account to exclude estimates when sufficient
data is available or use estimates when they are not.}
\usage{
imputeWithAndWithoutEstimates(data, processingParameters, formulaParameters,
  imputationParameters, minObsForEst = 5)
}
\arguments{
\item{minObsForEst}{The minimum number of observation required to keep
'E'stimated values.}

\item{meatKey}{The Datakey object}
}
\value{
Imputed time series
}
\description{
NOTE (Michael): This switching between imputation with/without manual 'E'
                stimates will eventually be removed.
}

