% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productionProcessingParameters.R
\name{productionProcessingParameters}
\alias{productionProcessingParameters}
\title{Default Processing Parameters}
\usage{
productionProcessingParameters(datasetConfig, dataset = "aproduction",
  removePriorImputation = FALSE, removeManualEstimation = FALSE,
  keepOnlyProtected = TRUE, imputationObservationFlag = "I",
  imputationMethodFlag = "e", balanceMethodFlag = "i",
  manualEstimationObservationFlag = "E", manualEstimationMethodFlag = "f",
  missingValueObservationFlag = "M", missingValueMethodFlag = "u")
}
\arguments{
\item{datasetConfig}{The dataset configuration returned by the function
\code{GetDatasetConfig}}

\item{dataset}{The dataset name in the database.}

\item{removePriorImputation}{logical, whether previous imputation values
should be removed.}

\item{removeManualEstimation}{logical, whether previous manual estimation
should be removed.}

\item{imputationObservationFlag}{The observation flag that represents
imputation.}

\item{imputationMethodFlag}{The observation flag that represents estimated
by statistical algorithm.}

\item{balanceMethodFlag}{The method flag that for calculation based on
identity}

\item{manualEstimationObservationFlag}{The observation status flag that
corresponds to manual (FAO) estimates}

\item{manualEstimationMethodFlag}{The method flag that corresponds to manual
(FAO) estimates}

\item{missingValueObservationFlag}{The observation flag corresponding to
missing values.}

\item{missingValueMethodFlag}{The method flag corresponding to missing
values.}
}
\value{
Returns a list of the default parameters used in the data
  pre-processing algorithm.
}
\description{
This function can be used to generate the input parameters for the data
pre-processing code.  This is a good way to get a list of the required
parameters and then modify parameters to match your particular
configuration.
}
\details{
Below is a description of the parameters: \itemize{

\item areaVar: The column corresponding to the geographic area.
\item yearVar: The column corresponding to the time dimension.
\item itemVar: The column corresponding to the item/commodity.
\item elementVar: The column corresponding to the measured element.
\item valueVar: The name for value column in the normalised form.
\item flagObservationVar: The name for observation flag in the normalised form.
\item flagMethod: The name for method flag in the normalised form.
\item removePriorImputation: Should previous imputations be removed?
\item removeManualEstimation: Should previous manual estimates be removed?
\item imputationObservationFlag: The flag for imputation.
\item imputationMethodFlag: The observation flag that represents estimated by statistical algorithm.
\item balanceMethodFlag: The method flag that for calculation based on identity
\item manualEstimationObservationFlag: The observation status flag that corresponds to manual (FAO) estimates
\item manualEstimationMethodFlag: The method flag that corresponds to manual (FAO) estimates
\item missingValueObservationFlag: The observation flag corresponding to missing values.
\item missingValueMethodFlag: The method flag corresponding to missing values.
}
}

