% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transferAnimalNumber.R
\name{transferAnimalNumber}
\alias{transferAnimalNumber}
\title{This function transfers the animal number element from the parent
commodity (e.g. cattle) to the animal slaughtered element of the
child commodity (e.g. cattle meat).}
\usage{
transferAnimalNumber(data, selectedMeat, areaVar = "geographicAreaM49",
  itemVar = "measuredItemCPC", elementVar = "measuredElement",
  yearVar = "timePointYears")
}
\arguments{
\item{data}{A data.table object containing the data.}

\item{selectedMeat}{The data table of the same format returned by
the function \code{getAnimalMeatMapping}, but containing the
meat selection required.}
}
\value{
A data.table object of the same dimension but with the
    values of the slaughtered animal updated by the animal number.
}
\description{
This procedure is to ensure the two element are in synchronisation
as required.
}

