% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transferParentToChild.R
\name{transferParentToChild}
\alias{transferParentToChild}
\title{Function to transfer values from parent to child.}
\usage{
transferParentToChild(parentData, childData, mappingTable,
  parentToChild = TRUE)
}
\arguments{
\item{parentData}{The animal slaughtered data from animal commodity.}

\item{childData}{The animal slaughtered data from meat commodity.}

\item{mappingTable}{The mapping table between the parent and the child.}

\item{parentToChild}{logical, if true, slaughtered animal are transferred
from animal commodity to meat, otherwise the otherway around.}
}
\value{
The transferred data
}
\description{
When values transferred between parent and child, shares are applied. To
transfer parent to children, we multiply by the share in order to obtain the
share of parents used to produce the child commodity. In the reversed
direction, the values are divided by the shares.
}

