% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useEstimateForTimeSeriesImputation.R
\name{useEstimateForTimeSeriesImputation}
\alias{useEstimateForTimeSeriesImputation}
\title{This function creates a table to determine whether a country and
element combinated time series should use include estimates in the
iputation process.}
\usage{
useEstimateForTimeSeriesImputation(data, areaObsFlagVar, yieldObsFlagVar,
  prodObsFlagVar, impFlags = c("I", "E"), missFlags = "M",
  minObsForEst = 5)
}
\arguments{
\item{data}{A data.table object containing the data to be imputed.}

\item{yieldObsFlagVar}{The column corresponding to the observation
flag of the yield.}

\item{prodObsFlagVar}{The column corresponding to the observation
flag of the production.}

\item{impFlags}{The value of the observation flags which are
considered to be imputed.}

\item{missFlag}{The value of the observation flag which
corresponds to missing value.}
}
\description{
TODO (Michael): This function should be expanded to include item as well.
}

